requirements.txt
rich
click
python-dotenv
pydantic
langchain
chromadb
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
requests